requirements.txt
numpy
python-mnist
matplotlib
pickle
PIL